Split dosimeter.py into 4 files
sensor.py, manager.py, sender.py, auxiliaries.py
